bug of no app_id

+++ access/delete.py
@@ -10,9 +10,9 @@ from . import db, User, Application
 @deletes.route('/delete/<int:app_id>')
 @login_required
 def delete_app(app_id):
-    if app_id == 0:
-        return redirect(url_for('access.tokens'))
     this_app = Application.query.filter_by(id = app_id).first()
+    if app_id == 0 or not this_app :
+        return redirect(url_for('access.tokens'))
     if this_app.user_id == current_user.id:
         db.session.delete(this_app)
         db.session.commit()
